lntest: move mining methods into one file
